refactor: rename PersistenceCache -> PersistentCache
also move SetDone class to its own file
