+++ src/main/java/io/retel/ariproxy/persistence/SetDone.java
@@ -1,0 +1,29 @@
+package io.retel.ariproxy.persistence;
+
+import java.io.Serializable;
+import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+
+class SetDone implements Serializable {
+
+	private final String key;
+	private final String value;
+
+	public SetDone(String key, String value) {
+		this.key = key;
+		this.value = value;
+	}
+
+	public String getKey() {
+		return key;
+	}
+
+	public String getValue() {
+		return value;
+	}
+
+	@Override
+	public String toString() {
+		return ReflectionToStringBuilder.toString(this, ToStringStyle.SHORT_PREFIX_STYLE);
+	}
+}
